ENH: break Matches pages into components
