+++ src/app/components/Matches/Match/Team/styled.js
