--- src/app/pages/Matches/Index/styled.js
@@ -1,4 +1,0 @@
-import styled from 'styled-components'
-
-export const Container = styled.div`
-`
